Individual-Graduation-Project/backend/requirements.txt
accelerate==0.34.2
aiosqlite==0.20.0
aiosmtplib==3.0.2
alembic==1.13.3
fastapi==0.115.0
httpx==0.27.2
passlib==1.7.4
pydantic==2.9.2
pydantic[email]
pydub==0.25.1
pytest==8.2.2
pytest-asyncio==0.24.0
pytest-ordering==0.6
python-dotenv==1.0.1
python_jose==3.3.0
python-multipart
SQLAlchemy==2.0.35
uvicorn==0.31.1
bcrypt==3.1.7
transformers==4.42.3
torch==2.4.1